Altera o tipo de dados de colunas de medição para REAL (#89)
- Em `GeOsDeviceItemScript.kt`, as seguintes colunas da tabela `ge_os_device_item` tiveram seu tipo alterado de `ColumnType.INT` para `ColumnType.REAL`:
    - `MEASURE_MIN`
    - `MEASURE_MAX`
    - `MEASURE_ALERT_MIN`
    - `MEASURE_ALERT_MAX`

+++ app/src/main/java/com/namoadigital/prj001/database/scripts/multi/ge_os/GeOsDeviceItemScript.kt
@@ -229,22 +229,22 @@ val geOsDeviceItemTable = with(GeOsDeviceItemDao) {
             ),
             DatabaseTable.Column(
                 name = MEASURE_MIN,
-                type = ColumnType.INT,
+                type = ColumnType.REAL,
                 isNullable = true,
             ),
             DatabaseTable.Column(
                 name = MEASURE_MAX,
-                type = ColumnType.INT,
+                type = ColumnType.REAL,
                 isNullable = true,
             ),
             DatabaseTable.Column(
                 name = MEASURE_ALERT_MIN,
-                type = ColumnType.INT,
+                type = ColumnType.REAL,
                 isNullable = true,
             ),
             DatabaseTable.Column(
                 name = MEASURE_ALERT_MAX,
-                type = ColumnType.INT,
+                type = ColumnType.REAL,
                 isNullable = true,
             ),
             DatabaseTable.Column(
